Add Photo supplier modal
