+++ client/components/Widgets/Modals/modal.module.css
@@ -1,7 +1,8 @@
 .container {
   position: relative;
   width: 30em;
-  height: 40em;
+  height: auto;
+  max-height: 40rem;
   background-color: rgb(255, 255, 255);
   font-family: Ubuntu;
   flex-direction: column;
