+++ client/pages/index.tsx
@@ -5,6 +5,7 @@ import Searchbox from 'components/Searchbox/Searchbox';
 import Modal from 'components/Widgets/Modals/Modify/Modify';
 import Info from 'components/Widgets/Modals/Info/Info';
 import dynamic from 'next/dynamic';
+import Pictures from 'components/Widgets/Modals/Pictures/Pictures';
 
 const DynamicMap = dynamic(() => import('../components/Map/Map'), {
   ssr: false,
@@ -18,6 +19,7 @@ const Main = () => {
       <Modal />
       <Info />
       <Navbar />
+      <Pictures />
       {searchVisible ? (
         <div className={modalStyle.app}>
           <Searchbox />
